Publish a player's current humanity rating in their GameMemberInfo so that we can compute the total available flow on a particular client and so that we can compute it from Whirled Java games where we don't know about MemberObject.

+++ src/java/com/threerings/msoy/data/MemberObject.java
@@ -202,9 +202,9 @@ public class MemberObject extends MsoyBodyObject
     /**
      * Return our assessment of how likely this member is to be human, in [0, 1).
      */
-    public double getHumanity ()
+    public float getHumanity ()
     {
-        return humanity / (double)MAX_HUMANITY;
+        return humanity / (float)MAX_HUMANITY;
     }
 
     // documentation inherited
